Account for Freshen_Cell_Header not testing raiseds
There was a check in Freshen_Cell_Header() at one time which didn't
let you freshen a cell if it contained a raised error.  This check
proved to be too arduous and removed, so if you have Init act as
a Sink then it no longer needs comments saying it interferes with
that check.