Fix bug when Level->out points into stale C stack cell
The way that longjmp() and throw is handled currently will jump up to
the Trampoline and give the last pushed Level* a chance to recover.

In the event that a C stack-based cell was used as a Level's ->out
pointer, then if the longjmp() or throw were to cross the definition
of that stack-based cell then the recovery/cleanup of that Level*
would be dealing with a corrupt pointer in Level->out.

This isn't a common problem, because there are relatively few stack
based cell evaluations--most write to another Level's out or spare
cell.  The few that exist generally are old stackful code that has
to run an invocation of the trampoline, and that invocation would
trap the longjmp() or throw before unwinding across the C stack cell.

So only a few cases have this problem--if they define a stack cell
and incur a failure between when a Level is pushed and a trampoline
is invoked.