PARSE-MATCH vs PARSE:MATCH, PARSE-THRU vs PARSE-
The PARSE- operation makes more sense with a name like PARSE-THRU

It's a bad idea to combine refinements like :MATCH and :THRU so having
PARSE-MATCH and PARSE-THRU be separate entry points makes more sense.
This resolves a bug in destructure.

PARSE-THRU as implemented here has a weakness of being dependent on
the ACCEPT and `<here>` combinators.  PARSE-MATCH is written to not
have that weakness.