New Generic Strategy
This is a major change to how generic functions are implemented, based
on the idea that you can IMPLEMENT_GENERIC() for a given type or
typeset, and the source is scanned for these definitions to make
tables that dispatch to those functions.  It's a complete replacement
for the large switch() statements based on symbol IDs.

The goal is that it will replace Makehooks, "To" hooks, Compare hooks,
Mold hooks, etc. as well as provide the ability for generics to have
better invariants and common code before and after the dispatch based
on the type to the appropriate IMPLEMENT_GENERIC() function.

In order to untangle the dependencies, the code from %make-boot.r that
pertained to building the type tables was migrated out into its own
script (%make-types.r), which produces a table that can be read for
just the amount of type information that make-boot.r requires.  (This
is also good because %make-boot.r had become huge and complex, so
cutting its size in half makes it more manageable.)