Accelerate built-in typesets (formerly "decider")
The original concept of packing bytes in the typechecker was that the type
check could be of anything, so it needed the full cell.  This means that
CHAR? as a type predicate could check that something was an ISSUE! but also
a single codepoint.

In practice this hadn't been gotten around to yet, as the table building
was just making functions that checked the fundamental type for being in
either a sparse typeset (checked via flags) or a ranged typeset, or a
fundamental datatype.

The move to try and use "deciders" in generics meant that the performance
would be worse if it wasn't able to just check a datatype against either
a range or a sparse set.  Given that fact, this backs off the idea that
the optimization is geared toward arbitrary checks and instead optimizes
the type range and sparse checks (where datatype checking is a trivial
case of a range that starts and ends at the same place).

While this is more limited in terms of what the optimized bytes can
represent, it creates a high-leverage tool that simplifies the code a
significant amount, and can be used to efficiently hook generics onto
"typesets" as well as fundamental types, indicated by a single byte.