Replace "CompareHook" with IMPLEMENT_GENERIC (WIP)
This takes a first step of wiping out the CT_Xxx functions by making
them called from generics.  It bases comparison on EQUAL? and LESSER?
which follows the pattern of Haskell's Ord and Eq classes, or C++
standard library performing sort based on operator< and operator=.

While this does eliminate the CompareHook table in favor of dispatch
through the new generics model, the CT_Xxx functions still exist and
will take some time to excise.  Because comparison will be based on
building frames for the EQUAL? and LESSER? functions, there will
have to be more care taken in terms of performance in places where
equality tests are done... such as in switch() statements which
for the moment go through the API to call EQUAL? as a predicate.