Replace "MoldHook" with IMPLEMENT_GENERIC (WIP)
This transitions to a model in which molding is done through calling
the evaluator.  It means building a frame for each new molded item,
which adds cost...but by doing so it means we're able to use the
generic molding mechanism.  This provides the ability to extend
molding in the future, and removes one more wacky dispatch.