Build old generic machinery on top of new generics
This redesigns the old generic machinery to use the new generics,
allowing for the complete removal of the "TypeHooks" concept.

Basically the whole switch() based generic mechanism is modeled
as a single "oldgeneric" native.  So it uses the new generic for
the type dispatch but then uses the switch().

The "verb" symbol is slipstreamed in by overwriting whatever the
label was that was used to call the function, which is good
enough for a temporary measure until all the oldgeneric are
killed off.