Use new generics for AS and TO
The introduction of AS and TO threw a wrench into the dispatch
because it made more sense to define TO methodization as being
done by the type converting, which is the second argument.

Now that generic dispatch has been rethought completely, the
issues that a lack of commonality for the switch() statement
caused based on the shifting position of the argument no longer
apply.  There's no switch, and instead of goto the system can
use GENERIC_CFUNC().