Make set operations (e.g. UNION) use new generics
Historical Rebol had an all-or-nothing generics handling idea,
where if you registered a handler for Binary then it would mean
all operations would go through the C function with a switch()
statement for Binary.  This meant that finer-grained handling or
passing on of events had to be done manually.

New generics don't have this problem.