Rearrange %sys-patch.h to stubs/stub-use.h
The %sys-patch.h file is a holdover from a time that may have been
before Flavor bytes, where "Patch" was used to handle the fledgling
LETs, USEs, and module variable instances.  But with flavors they
broke out into their own independent types.

Now "Patch" is part of the SeaOfVars include file, so include file
code for Use needs to be somewhere different.

This helps the push toward having better organization of the files
in include so there aren't so many sys-xxx.h files in the root
include directory.