Restore DID and DIDN'T to original meanings
The idea that THEN? and ELSE? as the prefix form of testing if a
branch had been taken was "too obscure" led to a period of thinking
that DID and DIDN'T, instead of being simple synonyms for LOGIC?,
should take over THEN? and ELSE?'s testing for "pure null or
pure void".

Reason was that the results from things like MATCH and PARSE were
too hard to manage with edge cases, and they were using pack
antiforms to get around ambiguity of synthesized null vs. pure.

However those two cases were changed: MATCH added a MATCH:META mode
and errors on pure null otherwise.  PARSE started giving raised
errors if it didn't complete or match.  So the common scenarios
got better ways of solving the problem, and it's best to push DID
and DIDN'T back to their original meaning.

Defines BRANCHED? as an alias for THEN? in case that makes it more
likely to be leveraged somewhere.