Eliminate REFLECT new virtual binding OF
After observing that the OF native has a binding environment
available, it was noticed that it could actually look up
a name derived from its left hand side.

This means you could write HEAD OF XXX and the OF operator
could actually look up HEAD-OF and run it through a reeval
to give you the same effect as HEAD-OF.

Combined with the new generics implementation, this lets the
names of the generic functions be things like HEAD-OF and
yet still get all the benefits of the "reflectors".

This goes ahead and changes all the reflector instances, but
doesn't have support for the old generics model passing through
to the "port hooks" yet.  This means ports only dispatch the
old-style generics.

Note: if REFLECT were desired, e.g. REFLECT XXX 'HEAD, it could
be achieved with the same mechanism used by OF.  However, OF
is effectively variadic since it reevaluates...meaning it
works with MINIMUM OF or other arity functions.  So REFLECT
would only work with arity-1 functions.