Redo casing for new generics to get SymId
This changes from:

    IMPLEMENT_GENERIC(multiply, integer)
    IMPLEMENT_GENERIC(append, any_list)

To:

    IMPLEMENT_GENERIC(MULTIPLY, Is_Integer)
    IMPLEMENT_GENERIC(APPEND, Any_List)

The reason for the change to the uppercase for the generic name is that
it means token pasting can prepend SYM_##name and get a valid symbol
like SYM_MULTIPLY or SYM_APPEND.  This means Dispatch_Generic has the
symbol it's dispatching (useful for errors, but also for bridging to
the legacy port model which still uses switch() statements)

Changing the type constraint to use mixed case and Is_ in the case of
single typechecks helps make it stand out, and be more searchable
because ", port)" occurs all over the code.  But ", Is_Port)" would
be specific to a generic.  Not only does it look like the type
constraint functions, but jump-to-decl will work on it because the
Any_Series() and Is_Port() types of definitions are looked up, so
you can even see the type byte ranges.