Give antiforms REB_XXX types for each category
This makes some changes to %types.r, removing QUOTED! and QUASIFORM!
from it, and adds in descriptions for the antiforms.  It then goes
on to make it possible for VAL_TYPE() to return bytes specific to
antiforms so that you can switch on an antiform's class type.

Several nice advantages arise, like being able to typecheck antiforms
in the early boot natives (e.g. NATIVE itself can say it returns an
ACTION?) and functions like ACTION?, TRIPWIRE? and SPLICE? are
intrinsics which are automatically made using the pattern of the
other typeset testers.

Currently this leaves a large gap in the typeset-bytes because the
method for calculating things like REB_SPLICE is just to add the
value of REB_QUOTED to the heart, to push it out of band.  Since
not all fundamentals are isotopic, this means a lot of bytes go
unused.  However, these could be reclaimed and filled in with
typesets as the values could never be returned from VAL_TYPE().

That would add more complexity, and the somewhat disorganized
code in %make-types.r should be simplified before it's undertaken.