Reorder symbols so XXX-OF occurs after XXX for builtins
This rethinks the symbol generation process so that it's more dynamic,
not pinning down symbols to IDs until all the symbols have been
gathered.  Using TAG! in the symbol table, it's possible to mark
positions and then translate those to finalized IDs after the full
rearrangement process has gone on.

With this being possible, it means we can actually reorder things
to get advantages.  So for instance, since LENGTH OF wants to be
able to find LENGTH-OF in order to look it up and run it, it's
convenient to move the symbols so that LENGTH-OF is immediately after
LENGTH...then for builtin symbols, you can just add 1 to what you
get instead of going through the trouble of creating a derived
UTF-8 byte buffer with "-of" added to the string...hashing and
interning that...and navigating to the symbol.  You just add one
to the ID and go directly to that slot in the global fixed array
to find the symbol.

It doesn't work for everything... symbols that have other stricter
ordering constraints will have a problem.  But so far it's only
SIGIL-OF and FILE-OF, and simple exceptions can be made for them
at nearly zero additional cost (runtime or maintenance)!