Change REB_MAX to inclusive convention
In C, there's a pattern of putting maximum values in-band of
enumerations, like:

    typedef enum {
        COLOR_RED,
        COLOR_BLUE,
        COLOR_GREEN,
        COLOR_PURPLE,
        COLOR_MAX
     } ColorEnum;

This way you can do things like declare arrays which have
an entry per color:

     Pencil pencils[COLOR_MAX];

However, it has a couple of side effects.  It means that
you've added an entry in the enum which isn't "real" so
checks that you're handling every enum element in a switch
statement require a default: that asserts or you have to
mention the _MAX entry explicitly.

There are other good arguments for an inclusive MAX, and
a very compelling one is that things like the C defines for
integer boundaries must use an inclusive max as you would
not be able to represent it in the integer.

This is a step toward systemically switching to inclusive
max across the board.  It does mean you have to do "+ 1"
in array bounds that use the max, but there are many other
benefits.