Implement COPY in new generic dispatch
This continues along with the grind of converting old generics to new.

One interesting aspect of COPY is that it uses the fallback method
so that instead of implementing trivial COPY handlers in every
immediate type it's done in the COPY native when there's no specific
handler registered.

(The question of whether trivial copying should be allowed with COPY
is an open one, but it's good to see the technique working.)