Error overhaul: no Atom*/Cell*, permit Symbol*/nullptr
Frequently when you wanted to pass a word to an Error maker, you'd
have a symbol in your hand.  But the error maker only took cells,
so you'd have to make a temporary cell, come up with a name for it,
Init_Word() with the symbol, and pass that.

That was designed before Detect_Rebol_Pointer().  Now that we have
it, there's no particularly good reason not to accept a Symbol* and
detect it, then make the WORD! automatically.

It's also convenient if nullptr can be passed to get a ~null~
antiform.  Null fields in errors are common, so it seems okay to
allow you to pass them.

Building on that, this moves to having the label field of errors
that are reporting functions that may be anonymous report them
as ~null~ antiforms instead of ~anonymous~ quasiforms.  It may
be a little less communicative in the error readout, but saying
the label is null is more actionable in terms of being able
to write `if not error.label [...]` instead of having to compare
against '~quasiform~ ... the benefits of saying ~anonymous~ over
~null~ aren't really worth it for the added complexity, when
there is that advantage.

In order to check that you're only passing stable Value* and
Symbol*, this adds that check to the C++ build.  That caught
many cases that were passing Cell or Atom.  This patches up
those cases.

Additionally, there were some places that were handling the
fact that antiforms didn't have types at various times.  Now
they do, so the error messages don't have as many edge cases
to deal with...any Value* can come back with a type.