Embrace antiform types, e.g. SPLICE! vs SPLICE?
When antiforms first came on the scene they all gave back the same
type... which was antiform!.  This lined up with QUOTED! and
QUASIFORM! which are containers that don't make distinctions on
the type.

New techniques and reworking has made it possible to have antiform
types be separated out, and there's no real disadvantage to using
them.  This goes ahead and replaces the antiforms in type specs
with the exclamation part form.

(It also takes edge cases in the internal code that used to warn
"you can't take VAL_TYPE() of an antiform" and starts removing
those, because now you can.)