Don't make HEAD and TAIL globally defined
It's tempting to use HEAD and TAIL as variable names.  With the
ability to write either HEAD-OF or HEAD OF in your source, it
makes more sense to leave them not defined, and people who need
the abbreviations can just say:

   head: head-of/
   tail: tail-of/

Either globally for their module or locally to a function that is
doing lots of `head tail back back head` kinds of things.