Overhaul quoting byte handling
Quoting and antiforms were designed before the implementation of the
cell subclasses (Element, Value, Atom), and also before the invention
of the contravariant Need() and Sink() containers.

This goes through and makes the quoting routines enforce their rules
more correctly, e.g. by making sure Quotify() takes in an Element*
and returns and Element*.  The updates make it so that you can't
accidentally wind up writing antiforms where you shouldn't.

It also introduces QuoteByte as a synonym for Byte to help show where
a Byte conveys the quote_byte information.

Additionally, since it was decided that not all types would have
antiforms and quasiforms, this changes (most of) the places that fail
on checking it to bubble up the error cooperatively, instead of
abruptly via fail() and longjmp().