Redesign RANDOM as part of making it a new generic
Historical Rebol had a very overloaded RANDOM function.  This
breaks it into several different functions:

     RANDOM:SEED => RANDOMIZE
     RANDOM:ONLY => RANDOM-PICK
     RANDOM => if it shuffled a series, SHUFFLE, else RANDOM

Additionally, the ability to get a shuffled version of immutable
values is added via SHUFFLE-OF.  This uses the same technique as
e.g. REVERSE-OF, which is to look if the type supports a specific
SHUFFLE-OF method and if not fall back on running SHUFFLE on a
COPY of the value.  Hence while ANY-SEQUENCE? and ANY-WORD? define
a SHUFFLE-OF handler, ANY-LIST? does not but you can still get a
fresh copy of a series.