Factor out AS generic behavior as C function
AS conversions don't do any evaluation (at least, not of the data
they are passed...they may be lazy and use the evaluator to do
some of their work).  So they factor cleanly into a C function
that doesn't return a complicated `Bounce`.  Those C functions
can be used various places that want to do AS conversions and
don't want to invoke the evaluator.