Restore COMPOSE to arity-1, add COMPOSE2
The concept of using the @ signal to mean "steal binding from the
pattern", along with the name COMPOSE2, makes it more palatable
to return compose to being arity-1.

Also, INTERPOLATE is planned to work on both blocks and strings
as single-arity which uses the () pattern bound in the evaluation
environment, thus meaning it is "sneaky" in a way compose never is.