Type_Of() returns Option(Type)
This lays groundwork for resurrecting "extension types"...which use a
reserved byte for the Cell's heart to indicate that the real type
information is in the Cell's extra slot.

One thing different in this resurrection is that the choice of the
heart byte to indicate an extension type is intentionally 0.  This
lets the tricky Option() C++ class document the places where the type
has either been determined to be builtin or not, and it can be tested
with a simple `if (type)` to know if it's built in or not.  The
original attempt to build extension types used a non-zero value and it
was a complete mess.