Extension FLAVOR_XXX, FLAVOR_BYTE=>TASTE_BYTE
There's a bit of a problem calling the byte in the series that stores
the "Flavor" a FLAVOR_BYTE.  Because that makes it look like a literal
for FLAVOR_XXX, as if it's one of the flavors.

This renames the byte to the TASTE_BYTE.  It's a strange name but you
don't see it that often, since most of the time you use Stub_Flavor()
to read it, or FLAG_FLAVOR() to produce the flag.

With this solved, we can adds a generic flavors that extensions can use,
of FLAVOR_BYTES, FLAVOR_CELLS, and FLAVOR_POINTERS.