rebUnmanage() just unmanages values
There was an idea that rebUnmanage() could work with either API handles or
with memory allocated through the API.  That is technically possible to
do...because both have information in the bytes before the pointer that
was handled out that could be used to distinguish them.

However, this is extra technical work for questionable advantage, vs. just
having two routines.  It's not a useful polymorphism, and it's more obvious
at the callsite what's going on with two routines.  Additionally they can
return the unmanaged value right back for chaining, as the right type, if
they are two entry points.