Include C89 aliases even in non-C89 librebol includes
If you're going to put #ifdef inside of the code you pass to something
like rebValue(), then you can't use the __VA_ARGS__ macro version.

Hence we should include the C89 versions always.  The difference in
the LIBREBOL_USE_C89 variant is that it *requires* you to use the C89
versions, by not defining the other ones.