Add placeholder API functions for rebLock/Unlock Bytes
It's become clear that it is desirable for extensions written to the
librebol API to be able to access the underlying bytes of binaries
and strings.  But there will have to be some care taken to lock
those series so they can't move under expansion or other behavior
when the evaluator is called.

The routines are needed now, so this gives the access minus the
necessary protection, and the extensions are trusted to not do
anything that will mess it up (and the GC doesn't do any improvised
compacting or relocations yet, so not setting a flag on the series
isn't a problem...yet).