Drop leading slash for FUNC, METHOD, LAMBDA, DOES
It's not clear exactly what the strategy is going to be for the
leading slash, but currently it seems like it's definitely going
to be optional--at least for assignments from "known function
generators" (e.g. functions that explicitly return ACTION! vs.
just kind of incidentally try to as part of returning ANY-VALUE?)

This drops the leading slash from the most common obvious
generators, with likely more leading slashes to be dropped.  But
having a mix helps test the feature for now.