DECLARE_NATIVE(xxx) => DECLARE_NATIVE(XXX)
All-Caps has better properties for converging SYM_XXX constants with
the native constants in C code.  It also visually helps distinguish
the macros from the variable names:

    Value* something = ARG(something);
    =>
    Value* something = ARG(SOMETHING);

So it's more clear that you're not passing a variable as a parameter.

This convention was adopted in the mainline, and it was easy enough
to bring the bootstrap executable into sync with it, which minimizes
friction going between them.