Option(Level*) for running frames, kill REDO
The REDO native has been dropped from modern Ren-C, so there's
certainly no reason to keep it alive in the bootstrap executable.

This also makes the handling of Level* clearer in terms of when
you are trying to get a Level for a frame that may or may not be
running.  The form that doesn't fail implicitly if it can't get
the level returns Option(Level*) instead of just Level*.