Kill off refinement promotion
This gets rid of a lot of convoluted code that had some bug in it, and
it was easier to delete it than to fix it.

The behaviors which it was trying to implement were based on the model
of refinements being separate from their arguments (and potentially
having multiple arguments)--none of that applies anymore, so there is
no need to do more than the minimum in specialization.