Make extended symbol and generic registration automatic
%prep-extension.r now goes ahead and hooks into the STARTUP* and
SHUTDOWN* functions the boilerplate work that's needed to register
generics, extension words, and a hacky way of registering the
extension types as well.

This means extensions no longer need to do this manually, so the
STARTUP* and SHUTDOWN* functions are empty in the extensions that
are defining custom datatypes at the moment.  This also keeps the
idea that you can omit those functions if you don't want them,
without interfering with the automatic behaviors.