Call what were "Decayed" stubs "Diminished" stubs
The idea of stub decay is to enter a state where on a later GC pass,
all references to it will be canonized into references to one single
global freed stub. This reclaims memory that freed stubs would take
up if the references were not canonized--providing no useful
information and taking up space.

This concept predates isotopes, so the term "decay" hadn't taken on
the strong meaning that it has now with unstable isotope decay.  This
changes the name to "Diminished" stubs.