Get rid of special cases in UintptrUnion
Trying to put every type into the UintptrUnion caused problems
with the dependencies of the header files, making things be
declared sooner than they otherwise would be.  This shifts the core
to use the same practice as extensions, to do their own casting of
the fundamental types in the UintptrUnion.

(It also renames AnyUnion to UintptrUnion, which makes more sense.)