Bring back typechecked Bounce concept
There was an idea to use a C++ struct to represent the Bounce concept
so as to filter out unwanted pointer types.  This was dropped when
API dispatchers that had to use void* showed that the struct was
not compatible.

But things have changed: API natives have a different spec and take
a RebolContext* and not a Level.  So they are distinct, and the core
natives can take the Bounce and be incompatible.  This is worth it
for the extra checking--and in fact it caught several mistakes of
returning Error* instead of FAIL(Error*) or RAISE(Error*)