Crypto extension cleanup, use rebBytes
This revisits the crypto module as part of trying to get rid of uses
of "blockscope {...}" which is ugly compared to the new use of goto
labels and blocks to mark sections.

While doing the update, it switches over to rebLockBytes() instead
of using rebBytes() so as to not make copies of data being encrypted
or encrypted.  It also moves to only call librebol cleanup routines
on non-erroring paths, to help emphasize that those cleanups are
automatic if a failure occurs (code above depends on it, so why not
the code at the end of the routine?)