Enable goto out of RESCUE_SCOPE_XXX
It wasn't possible to goto because the longjmp() trickery version
was initializing a variable based on the declaration.

This sacrifices the syntax looking more like C++ in order to remove
an illusory degree of freedom of setting the type of the error
declaration, for the purpose of enabling gotos out of the scope.