Add ADDRESS-OF generic
Previously, ADDR-OF was a fixed function that only understood STRUCT!
and ROUTINE! and was part of the FFI extension.

This enables other types to answer the question, and makes it the
mechanism that VECTOR! uses to answer it for the FFI to use, so that
vector is decoupled from the C implementation.