Use MAYBE instead of UNWRAP in type macros
The parameters suggest that Option(Type) is legitimate.  So when instances
of Option(Type) come along, that gets checked.

(This may need review, e.g. object subtypes might answer ANY-CONTEXT? but
be TYPE_0 for custom classes.  But so long as it takes Option(Type) it
needs to use maybe instead of unwrap.)