Permit item-wise JOINs with non-list/sequence
There's some concern about what it might mean to join with a
list or sequence.  But it seems that if you say:

   join 'abc 1

It's clear that should be abc1 with no ambiguity.  So allow such
joins while not allowing:

   join [a b c] 1

Due to the slippery slope when that becomes:

   join [a b c] [1]

(And you might have this obfuscated by variables.)