Allow UUID module to build on mac
The inclusion of a definition of Size in %sys-core.h means you can't
include both `<CoreFoundation/CFUUID.h>` and sys-core.h

This isn't a problem in modern Ren-C due to the UUID module using
librebol.  But that's not an option in this old bootstrap build, so
just move the UUID-getting functionality out into its own file.