Some Rebmake COMMAND -> COMPILE and LINK
In Rebmake, COMMAND was a function that had different parameterization
based on whether it was just requesting something to make its text
line in a makefile (e.g. arity-0), or it was the name of a function
parameterized to either compile or link.

This polymorphism of what COMMAND meant was confusing and not useful,
and also inhbits the idea of rigging up something like GCC as a
tool that can both compile and link.