Simplify to use C front end to link
Rebmake was originally written with a concept that you could specify
the linker independently of the compiler.  In practice, trying to link
without the C front end doesn't work very well... and if you really
need to you can ask the C front end to pass flags to the linker.  This
simplifies matters by cutting all that and letting the compiler do
the link step.

(This also gets rid of the separate linker functions for LLVM-LINK,
as clang's front end lines up with gcc's)