Fix problems in %mod-javascript.c
emcc was previously unable to use the RIGOROUS mode due to Rebmake
not distinguishing GNU-only warnings from those in clang, but now
that that difference exists it's possible to build mod-javascript.c
at higher warning levels and catch them as errors.

Also, technically speaking being able to have `...` variadic slots
in macros be empty is a C++20 feature.  Clang is pedantic about it.
We could disable the warning or just use a parameter where it isn't
really needed, and be C++11 compliant.

The use of "$" inside the macros is a "wontfix" issue from emscripten
so disabling warnings pertaining to it is the only option.