rebInteger() vs rebInteger64()
Emscripten flipped a switch where the int64_t type would be treated
in the ccall/cwrap bindings as a BigInt in JavaScript.

BigInt is awkward to work with and has no implicit conversions to or
from JavaScript's "Number" type.  This is unwieldy and so it seems
the only real course of action is to make separate functions for
handling bigger range integers.