Prohibit API calls during handle cleaners
There was an idea that you could do API calls specifically for things
related to handles during GC, but that's not right because the APIs
could do arbitrary evaluations.  However, API calls had to be used to
extract the handle pointer from the value.

This changes it so the handle cleaner function gets the extracted
properties, hence API calls aren't needed.