Switch LIBREBOL_USES_API_TABLE to #if vs #ifdef
As a general policy, Ren-C uses defines of 0 and 1 instead of just
"defined or not" as a protection against typos.

This brings LIBREBOL_USES_API_TABLE into the fold for that, and cleans
out the REB_EXT and REB_HOST options, which are unclear and don't
reflect the true orthogonality of choices.

Also fixes LOAD-EXTENSION for dynamic libraries in the process.