enum implements as uint_fast8_t, not int_fast8_t
Compilers can make the int_fast8_t any size integer, which means they
may have the behavior that assigning 255 to them would be signed or
unsigned.  (Seems to be signed on Mac, Linux, Windows... but winds up
being unsigned on HaikuOS)

There was no reason to be using a signed quantity to begin with, but
being signed masked a buggy issue in Copy_Cell_Header() where 255
was comparing underneath TYPE_NULLED instead of above it.