Rename NOTHING => TRASH
The initial name of antiform blank was "trash".  It was deemed too much
of a "reified" name to use, and then taken for ~ quasiforms.

This has been deemed misguided.  It put the name NOTHING as too close
to VOID.  The lack of reification of "TRASH" is not that important...
all antiforms lack reification.