Allow POKE of TRASH, split into POKE/POKE*
POKE previously did not take a ^META parameter, which meant it could
not pass through raised errors.

This transitions POKE into two operations: POKE which returns the
value you passed in, and POKE* that has the "give me the writeback"
character.  If an operation gets to the top level POKE and requires
writeback, it's an immediate value and cannot be poked.