Add ANY-ELEMENT! typeset
There was a lot of confusion prior to antiforms on what was meant by
"ANY-VALUE!" in terms of whether it included the "NULLVOID" of that
era.  Today with antiforms there's a category of "element" which
covers everything you can put in a block, while "value" is everything
you can put in a variable.