Drop MKF_KEYWORDS, MKF_FAKE_RETURN
The MKF_FAKE_RETURN optimization was part of an early attempt at what
are now "intrinsics" in the main branch.  It really only generates
differences between the non-debug and debug builds, so just causing
potential problems with that vs. being a great performance advantage.

MKF_KEYWORDS was applicable when MAKE ACTION! was a raw action
generator.  New methodologies in the main branch use frames to get
the "no keywords" nature of function synthesis.