Transition to "Flexible Logic"
A tricky idea here is to introduce ~okay~ antiforms as a new type,
and then LOGIC! becomes a tpyeset unifying ~okay~ and ~null~.

This builds on that, as well as changes to using trash as the
unspecified value in frames, which makes it easier for ~null~ to
carry information.

In order to make progress on this, it required a stronger notion of
META and UNMETA than the "UNEVAL" hack offered.  So this just bites
the bullet and considers the creation of the GROUP! or WORD! to be
"META".