Drop invisibility mechanic in bootstrap
The mechanic has changed completely, and there's nothing to learn from
debugging problems with the old way.

Having ELIDE and COMMENT produce VOID handles many cases.

(This is actually part of a broader attempt to simplify the bootstrap
executable to make more common behaviors, like infix, work the same
way in the bootstrap exe as mainline.)