Simplify/fix infix deferring mechanism
The bootstrap EXE still had a very preliminary implementation of infix
deferral, that made it much more complex than it ultimately needed to
be.  Basically, things that intend to be put in places like the
conditions of an IF (such as AND, OR) should not use infix deferral,
and the weird mechanisms for enabling it were misguided.

If you limit infix deferral to the behavior of things like THEN and
ELSE, you can simply have a flag that is shared in common on an
evaluation and its child levels.