AND/OR/XOR => logic! result, group RHS
There was an idea that AND/OR/XOR would preserve truthy or falsey
values, but this is no longer interesting as ELSE and THEN effectively
react to the only falsey value (null).  Also, an idea was that a
GROUP! would be always evaluated while a BLOCK! would be evaluated
in a short-circuit way, but this is not obvious to the reader and
not a very interesting feature since you always want to short circuit,
and GROUP!s look better.