Make Cell_Crumb() part of heart byte
This demonstrates the ability to hold 1 of 4 states in the HEART_BYTE()

It means the TYPE_SPECIFIC_A and TYPE_SPECIFIC_B flags are free, e.g.
to indicate purity on functions.

(The goal of this was so that the crumb in the heart could hold the
sigil of the first slot of sequences, as a compression.)