Simplify API Run_Va() functions
This clarifies Run_Va() by using flags that are specific RUN_VA_FLAGS,
vs. passing in arbitrary evaluator flags.  It cuts down on the entry
points and makes things clearer.

Also this removes a TRAMPOLINE_FLAG_META_RESULT instance.  The new
idea is that this will not be a service of the trampoline, but rather
that the stepper evaluator will specifically only speak in meta results
in order to distinguish the "no result" case as a non-meta trash.