Make Stepper_Executor() always ^META
This gets rid of the idea of LEVEL_FLAG_META_RESULT, as a generic tool
for asking for a meta evaluation from a level...and instead takes the
main executor that wants to speak in meta terms and makes it always
speak in meta terms.

There's an important reason for Stepper_Executor() to speak in a meta
protocol (besides the fact that all frame arguments are being switched
to meta protocol by default).  Some callers are interested to know if
the stepper hit the end of the output, and this shouldn't be tested
for in advance.  e.g. EVAL:STEP [, , ,] shouldn't have to look at a
block itself and see that it's all commas--it should come back as
a result of the step.