Make LAMBDA able to vanish
This means if you write (lambda [] [comment "hi"]) it makes a vanishing
function.

Without this there'd be no way to return a GHOST! from a lambda, which
seems limiting for this fundamental part.  It's also inconsistent with
other fundamentals like parse rule blocks, in terms of representing
the ability to take something you could have written in place and
then factor that thing out with fully equivalent behavior.