Make COMMENT, ELIDE, etc. native intrinsics
There was a good reason early on to make the basic invisible functions
usermode, because they were tricky specialized infix functions and it
was necessary to show users they could make them too.

In the isotopic era with GHOST!, it's trivial for users to make
vanishing functions and work with the meta-results.  So might as well
go ahead and make them fast intrinsics.