Transition the `<maybe>` tag to `<opt-out>`
Under the new concept for MAYBE and OPT, it is OPT(IONAL) which turns
nulls into voids.  Hence it's beginning to make sense to refer to
a parameter that accepts void as being "optional".

The name "opt-out" makes perfect sense here, as in "if optional, then
don't run the function and return null".

It looks better with a hyphen than it does without.