Scan abc/def.ghi as abc/def/ghi (no exception)
This gets rid of the quirky attempt to not fully transform TUPLE!
into path.  The idea was to make it easier to represent FILE! as
a PATH!, but it made it harder... since def.ghi still made def/ghi,
so you couldn't just assume paths that didn't end in blank had their
extensions in the last slot.