Introduce `<undo-opt>` to make void args null
This simplifies cases in usermode code where people want to avoid the
hassle of meta parameters just because they want to be able to get
voids.  If nulls aren't of interest, then turning the voids into
nulls makes it easier to process and the parameter does not have
to be meta.