WHEN => COND, :(...) => INLINE, VETO in parse
The IF in R3-Alpha PARSE was arity-1 and weird, but a necessary thing
to convert logic from the true/false domain into parse match/mismatch.
It was called WHEN for a bit but COND is a better choice.

The obsession with trying to decorate groups to inline them came from
an era where plain groups erased their results.  Now that they don't
you can just say [inline (...)] and be done with it.